SOLR-14917: Move DOMUtil and PropertiesUtil to SolrJ (#1953)
