+++ solr/core/src/java/org/apache/solr/core/PluginInfo.java
@@ -24,8 +24,8 @@ import java.util.List;
 import java.util.Map;
 
 import org.apache.solr.common.MapSerializable;
+import org.apache.solr.common.util.DOMUtil;
 import org.apache.solr.common.util.NamedList;
-import org.apache.solr.util.DOMUtil;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
