+++ solr/solrj/src/java/org/apache/solr/common/util/PropertiesUtil.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.solr.util;
+package org.apache.solr.common.util;
 
 import org.apache.solr.common.SolrException;
